Added comments what concurrency issue the class contains Added Simplest actor to show the most basic ones.

+++ src/main/java/cz/jskrabal/concurrency/plain/UnsafeBankAccount.java
@@ -4,8 +4,9 @@ package cz.jskrabal.concurrency.plain;
  * Created by Jan Skrabal skrabalja@gmail.com
  */
 public class UnsafeBankAccount implements BankAccount {
-    private int balance;
+    private long balance;
 
+    //Warning: Not thread safe! (modification of long is not atomic)
     @Override
     public void deposit(long amount) {
         if (amount > 0) {
@@ -15,6 +16,7 @@ public class UnsafeBankAccount implements BankAccount {
         }
     }
 
+    //Warning: Not thread safe! (more than one thread can pass the condition at a time)
     @Override
     public void withdraw(long amount) {
         if (balance >= amount) {
@@ -24,12 +26,13 @@ public class UnsafeBankAccount implements BankAccount {
         }
     }
 
+    //Warning: Not thread safe! (Value can be accessed during modification and method can return invalid number)
     @Override
     public long getCurrentBalance() {
         return balance;
     }
 
-    /* Warning: Not Thread safe! */
+    //Warning: Not Thread safe! (Obvious reasons :-))
     @Override
     public void transferTo(long amount, BankAccount to) {
         withdraw(amount);
